requirements.txt
asgiref==3.5.0
astroid==2.11.1
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dill==0.3.4
Django==4.0.3
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.3
isort==5.10.1
lazy-object-proxy==1.7.1
liqpay-sdk-python3==1.0.3
Markdown==3.3.6
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
pylint==2.13.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
zipp==3.7.0
django-ckeditor==6.2.0
django-embed-video==1.4.2
django-js-asset==2.0.0
tzdata==2022.1








